refactor: 모달 댓글 리스트 컴포넌트 분할
